Move event loop runner to runner module
